Added what I have so far
This compiles but I get a segmentation fault when it tries to run line 163. I've tried adjusting code all over the place, which is why a lot of it is overly formal and redundant... trying to narrow possibilities. I'm going to his office hours this morning to try and figure it out + clear up some things regarding instructions